rename any4 file to quantize
